+++ src/main/java/epam/jwd/mobileTariffs/repository/TariffRepository.java
@@ -1,9 +1,9 @@
-package repository;
+package epam.jwd.mobileTariffs.repository;
 
-import mobileTarrifs.Tariff;
-import mobileTarrifs.ForeignBusiness;
-import mobileTarrifs.ForeignBusinessPlus;
-import mobileTarrifs.Success;
+import epam.jwd.mobileTariffs.mobileTarrifs.Tariff;
+import epam.jwd.mobileTariffs.mobileTarrifs.ForeignBusiness;
+import epam.jwd.mobileTariffs.mobileTarrifs.ForeignBusinessPlus;
+import epam.jwd.mobileTariffs.mobileTarrifs.Success;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -15,7 +15,7 @@ import java.util.List;
  * @version - 1.0
  */
 
-public interface TariffRepo {
+public interface TariffRepository {
     List<Tariff> TARIFF_LIST = new ArrayList<>(Arrays.asList(new Success(15.0, 43567, 500, 2000),
             new ForeignBusiness(28.50, 15687, 600, 2500, 300),
             new ForeignBusinessPlus(35.0, 9874, 700, 3000, 450, 30)));
